Add 'liked' and 'unliked' classes to like links
This is part of my master plan to bring back the [dig icon](http://fontawesome.io/icon/hand-peace-o/) from the old patchwork-classic. It's a styling choice to have such an icon, so we need a class that lets us know:

1. This is a like button
2. Whether the thing is liked or not.

Then we can add any icons we want in CSS.

I've never used Mutant before, so this should be inspected by someone who has.